--- packages/runtime/src/routes.ts
@@ -1,16 +1,0 @@
-import type { RouteItem, RouteModules } from './types';
-
-export async function loadRouteModule(route: RouteItem, routeModulesCache: RouteModules) {
-  const { id, load } = route;
-  if (id in routeModulesCache) {
-    return routeModulesCache[id];
-  }
-
-  try {
-    const routeModule = await load();
-    routeModulesCache[id] = routeModule;
-    return routeModule;
-  } catch (error) {
-    window.location.reload();
-  }
-}
